split the big __init__ into modules
